front end encore

+++ star-wars/src/Controller/UserController.php
@@ -18,6 +18,12 @@ class UserController extends AbstractController
     #[Route('/', name: 'user_index')]
     public function index(EntityManagerInterface $entityManager): Response
     {
+        // Vérification de connexion
+        if (!$this->getUser()) {
+            $this->addFlash('error', 'Vous devez être connecté pour voir la liste des utilisateurs.');
+            return $this->redirectToRoute('product_list');
+        }
+        
         $users = $entityManager->getRepository(User::class)->findAll();
 
         return $this->render('user/index.html.twig', [
@@ -28,6 +34,8 @@ class UserController extends AbstractController
     #[Route('/new', name: 'user_new')]
     public function new(Request $request, EntityManagerInterface $entityManager, UserPasswordHasherInterface $passwordHasher): Response
     {
+        
+        
         $user = new User();
 
         $form = $this->createForm(UserType::class, $user, [
@@ -71,6 +79,12 @@ class UserController extends AbstractController
     #[Route('/{id}', name: 'user_show', requirements: ['id' => '\d+'])]
     public function show(User $user): Response
     {
+        // Vérification de connexion
+        if (!$this->getUser()) {
+            $this->addFlash('error', 'Vous devez être connecté pour voir les détails d\'un utilisateur.');
+            return $this->redirectToRoute('product_list');
+        }
+        
         return $this->render('user/show.html.twig', [
             'user' => $user,
         ]);
@@ -79,6 +93,12 @@ class UserController extends AbstractController
     #[Route('/{id}/edit', name: 'user_edit', requirements: ['id' => '\d+'])]
     public function edit(Request $request, User $user, EntityManagerInterface $entityManager, UserPasswordHasherInterface $passwordHasher): Response
     {
+        // Vérification de connexion
+        if (!$this->getUser()) {
+            $this->addFlash('error', 'Vous devez être connecté pour modifier un utilisateur.');
+            return $this->redirectToRoute('product_list');
+        }
+        
         // Pour pré-remplir le champ roles, on récupère le premier rôle (ou ROLE_USER par défaut)
         $form = $this->createForm(UserType::class, $user, [
             'is_edit' => true,
@@ -118,6 +138,12 @@ class UserController extends AbstractController
     #[Route('/{id}/delete', name: 'user_delete', methods: ['POST'], requirements: ['id' => '\d+'])]
     public function delete(Request $request, User $user, EntityManagerInterface $entityManager): Response
     {
+        // Vérification de connexion
+        if (!$this->getUser()) {
+            $this->addFlash('error', 'Vous devez être connecté pour supprimer un utilisateur.');
+            return $this->redirectToRoute('product_list');
+        }
+        
         // La vérification CSRF est recommandée
         if ($this->isCsrfTokenValid('delete' . $user->getId(), $request->request->get('_token'))) {
             $entityManager->remove($user);
